+++ infobip-openapi-mcp-spring-boot-starter/src/main/java/com/infobip/openapi/mcp/autoconfiguration/OpenApiMcpConfiguration.java
@@ -4,6 +4,8 @@ import static com.infobip.openapi.mcp.autoconfiguration.Qualifiers.TOOL_HANDLER_
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.infobip.openapi.mcp.McpRequestContextFactory;
+import com.infobip.openapi.mcp.auth.CredentialProvider;
+import com.infobip.openapi.mcp.auth.HttpServletRequestCredentialProvider;
 import com.infobip.openapi.mcp.auth.scope.ScopeDiscoveryService;
 import com.infobip.openapi.mcp.config.ApiBaseUrlConfig;
 import com.infobip.openapi.mcp.config.ApiBaseUrlProvider;
@@ -146,6 +148,12 @@ class OpenApiMcpConfiguration {
     }
 
     @Bean
+    @ConditionalOnMissingBean
+    public CredentialProvider credentialProvider() {
+        return new HttpServletRequestCredentialProvider();
+    }
+
+    @Bean
     @ConditionalOnBean(MeterRegistry.class)
     public MetricService micrometerMetricService(MeterRegistry meterRegistry, NamingStrategy namingStrategy) {
         return new MicrometerMetricService(meterRegistry, namingStrategy);
@@ -168,8 +176,10 @@ class OpenApiMcpConfiguration {
             ErrorModelWriter errorModelWriter,
             OpenApiMcpProperties properties,
             ApiRequestEnricherChain enricherChain,
-            MetricService metricService) {
-        return new ToolHandler(restClient, errorModelWriter, properties, enricherChain, metricService);
+            MetricService metricService,
+            CredentialProvider credentialProvider) {
+        return new ToolHandler(
+                restClient, errorModelWriter, properties, enricherChain, metricService, credentialProvider);
     }
 
     @Bean